make susants website working

+++ core/service/jadmin/mfg/waxsetting.cpp
@@ -50,7 +50,7 @@ json WaxSetting::ins(json event, json args)
 {
     return insBase(event, args, "date, department_id, employee_id, description, status_id", "$1, $2, $3, $4, $5",
                    args[0]["date"].get<std::string>(),
-                                  args[0]["department_id"].get<int>(), args[0]["employee_id"].get<int>(), args[0]["description"].get<std::string>(),
+                                  args[0]["department_id"].get<long>(), args[0]["employee_id"].get<long>(), args[0]["description"].get<std::string>(),
                                   args[0]["status_id"].get<std::string>()
             );
 }
@@ -59,7 +59,7 @@ json WaxSetting::upd(json event, json args)
 {
     return updBase(event, args, "date, department_id, employee_id, description, status_id", "$1, $2, $3, $4, $5",
                    args[0]["date"].get<std::string>(),
-                                  args[0]["department_id"].get<int>(), args[0]["employee_id"].get<int>(), args[0]["description"].get<std::string>(),
+                                  args[0]["department_id"].get<long>(), args[0]["employee_id"].get<long>(), args[0]["description"].get<std::string>(),
                                   args[0]["status_id"].get<std::string>()
             );
 }
